Fixing the screen issues

+++ Smart_load_management_system/include/I2C.h
@@ -29,7 +29,7 @@ struct SensorData
   float batteryVoltage;
   float temperature;
   float totalLoad;
-  float percentage;
+  u_int8_t percentage;
 };
 
 extern SensorData data;
