+++ Smart_load_management_system/src/I2C.cpp
@@ -25,8 +25,9 @@ struct SensorData
   float batteryVoltage;
   float temperature;
   float totalLoad;
-  float percentage;
+  u_int8_t percentage;
 };
+
 
 SensorData data;
 void sendData()
@@ -37,39 +38,4 @@ void sendData()
 
   Serial.println("Data Sent");
 }
-// void setup()
-// {
-//   Serial.begin(115200);
-
-//   Wire.begin(); // Master mode
-// }
-
-// void loop()
-// {
-//   // Example values (replace with real sensor readings)
-//   data.energy1 = 10.5;
-//   data.energy2 = 11.2;
-//   data.energy3 = 12.1;
-
-//   data.voltage1 = 220.5;
-//   data.voltage2 = 221.1;
-//   data.voltage3 = 219.8;
-
-//   data.current1 = 5.2;
-//   data.current2 = 4.8;
-//   data.current3 = 6.1;
-
-//   data.frequency = 50.0;
-
-//   data.power1 = 1150;
-//   data.power2 = 1100;
-//   data.power3 = 1200;
-
-//   data.batteryVoltage = 27.5;
-//   data.temperature = 32.4;
-//   data.totalLoad = 3450;
-
-//   sendData();
-
-// }
